# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server API to control animalia kingdom",
        "title": "Animalia API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eduardo Freire",
            "url": "https://www.linkedin.com/in/edudufreire/",
            "email": "eduardofreire1995@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "paths": {
        "/citizens": {
            "get": {
                "description": "Find all citizens by filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Find all citizens",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Eduardo",
                        "description": "Citizen name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "cat",
                        "description": "Species",
                        "name": "species",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Happy Feline",
                        "description": "Description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Has pet human",
                        "name": "has_pet_human",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "id",
                        "description": "Sort by",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Number page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit rows by page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationCitizenBodyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new citizens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Create citizens",
                "parameters": [
                    {
                        "description": "Citizen data",
                        "name": "citizen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCitizenBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CitizenBodyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/citizens/{citizen_id}": {
            "get": {
                "description": "Find citizens by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Find citizens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Citizen ID",
                        "name": "citizen_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CitizenBodyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update citizen by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Update citizens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Citizen ID",
                        "name": "citizen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Citizen data",
                        "name": "citizen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCitizenBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CitizenBodyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete citizen by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Delete citizens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Citizen ID",
                        "name": "citizen_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.CitizenBodyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CitizenBodyResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "has_pet_human": {
                    "type": "boolean"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RoleBodyResponse"
                    }
                },
                "species": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.CreateCitizenBodyRequest": {
            "type": "object",
            "required": [
                "description",
                "height",
                "name",
                "photo_url",
                "species",
                "weight"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "has_pet_human": {
                    "type": "boolean"
                },
                "height": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.PaginationCitizenBodyResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CitizenBodyResponse"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "dto.RoleBodyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateCitizenBodyRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "has_pet_human": {
                    "type": "boolean"
                },
                "height": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}